\name{fmainPos}
\alias{fmainPos}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Calculate coordinates for title in plots
}
\description{
Internal function called by \code{plot.satin} and \code{quiver}.
}
\usage{
fmainPos(pu, main.pos)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{pu}{
Obtained by par("usr").
}
  \item{main.pos}{ Desired position, one among "topright", "topleft", 
                   "bottomright", "bottomleft").
}
}

\author{
Héctor Villalobos
}

