% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_ncvars.R
\name{name_ncvars}
\alias{name_ncvars}
\title{Get names of variables in a netCDF file}
\usage{
name_ncvars(nc)
}
\arguments{
\item{nc}{name of a netCDF file.}
}
\value{
This function only displays the number of variables, their names,
and corresponding long names.
}
\description{
This function returns names and long names of the variables in a netCDF file.
}
\details{
This function provides a practical way to know which variables are
contained in a netCDF file before importing the data.
}
\examples{
if(interactive()){
  name_ncvars("global-reanalysis-phy-001-030-monthly_1553098099635.nc")
}

}
\author{
Héctor Villalobos.
}
